backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
email-validator

# Зависимости для аутентификации
passlib[bcrypt]
python-jose[cryptography]
python-multipart
Pillow

# Зависимости для тестирования
pytest
httpx
websockets  # <-- ДОБАВИТЬ ЭТУ СТРОКУ